Extract DefaultIconProvider.
